requirements.txt
supabase==2.0.0
requests==2.31.0
python-dotenv==1.0.0
insightface==0.7.3
onnxruntime==1.17.0
opencv-python==4.9.0.80
numpy==1.26.4
scikit-learn>=1.3.0 
cryptography==44.0.2